+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Battlesnake
 
+## 0.1.5 (2023-01-04)
+
+### Changed
+
+  - Board
+    - #find_path(from, to, max_distance: nil) skips recursion if manhattan distance is greater than
+      max_distance.
+
 ## 0.1.4 (2023-01-03)
 
 ### Added